requirements.txt
Django==3.1.1
dj-rest-auth==1.1.2
django-allauth==0.42.0
django-cors-headers==3.5.0
djangorestframework==3.11.1
channels==2.4.0
channels-redis==3.1.0
mysqlclient==2.0.1
PyYAML==5.3.1
uritemplate==3.0.1 
nltk==3.5
numpy==1.19.2
pytesseract==0.3.6
uvicorn==0.11.8
uWSGI==2.0.19.1